requirements.txt
Django==4.2.16
psycopg2-binary==2.9.6
dj-database-url==1.2.0
whitenoise==6.4.0
python-dotenv==1.0.0
Pillow==10.0.0
django-cors-headers==3.13.0
django-simple-history==3.11.0
requests==2.32.3
gunicorn==20.1.0  # THIS LINE IS MISSING